+++ src/index.ts
@@ -1,3 +1,5 @@
+export { Client } from 'elasticsearch';
+
 export * from './lib/elasticsearch';
 export * from './decorators/index.decorator';
 export * from './decorators/primary.decorator';